==> old_notebooks/03_create_merged_mag_gtdb_table.ipynb <==
# 03_create_merged_mag_gtdb_table

Run this notebook to create a merged data frame of both the archaea and bacteria for all samples.

# load packages
import pandas as pd
import os
import os.path as op
import sys
import csv
import numpy as np
import matplotlib
import glob
import seaborn as sns
from collections import Counter
import matplotlib.pyplot as plt

# define functions 
def split_classification(df):
    df[['domain','phyla','class','order','family','genus','species']] = df.classification.str.split(';', expand=True)
    df['domain'] = df['domain'].str.replace('d__', '')
    df['phyla'] = df['phyla'].str.replace('p__', '')
    df['class'] = df['class'].str.replace('c__', '')
    df['order'] = df['order'].str.replace('o__', '')
    df['family'] = df['family'].str.replace('f__', '')
    df['genus'] = df['genus'].str.replace('g__', '')
    df['species'] = df['species'].str.replace('s__', '')
    return df

# file paths
path = '/Users/melissaherring/Google Drive/My Drive/MH_project/'
#path = '/Users/juliabrown/Google Drive/My Drive/projects/OMZvir_round2/MH_project/'

# input gtdb files
jv119_arc = pd.read_csv(op.join(path,'data/mag_data/jv-119_gtdbtk/jv-119.ar53.summary.tsv'),sep='\t')
jv119_bac = pd.read_csv(op.join(path,'data/mag_data/jv-119_gtdbtk/jv-119.bac120.summary.tsv'),sep='\t')
jv121_arc = pd.read_csv(op.join(path,'data/mag_data/jv-121_gtdbtk/jv-121.ar53.summary.tsv'),sep='\t')
jv121_bac = pd.read_csv(op.join(path,'data/mag_data/jv-121_gtdbtk/jv-121.bac120.summary.tsv'),sep='\t')
jv132_arc = pd.read_csv(op.join(path,'data/mag_data/jv-132_gtdbtk/jv-132.ar53.summary.tsv'),sep='\t')
jv132_bac = pd.read_csv(op.join(path,'data/mag_data/jv-132_gtdbtk/jv-132.bac120.summary.tsv'),sep='\t')
jv154_arc = pd.read_csv(op.join(path,'data/mag_data/jv-154_gtdbtk/jv-154.ar53.summary.tsv'),sep='\t')
jv154_bac = pd.read_csv(op.join(path,'data/mag_data/jv-154_gtdbtk/jv-154.bac120.summary.tsv'),sep='\t')

# create sample_name column
jv119_arc['sample_name'] = "JV119"
jv119_bac['sample_name'] = "JV119"
jv121_arc['sample_name'] = "JV121"
jv121_bac['sample_name'] = "JV121"
jv132_arc['sample_name'] = "JV132"
jv132_bac['sample_name'] = "JV132"
jv154_arc['sample_name'] = "JV154"
jv154_bac['sample_name'] = "JV154"

# create sample_depth column
jv119_arc['sample_depth'] = 400
jv119_bac['sample_depth'] = 400
jv121_arc['sample_depth'] = 95
jv121_bac['sample_depth'] = 95
jv132_arc['sample_depth'] = 80
jv132_bac['sample_depth'] = 80
jv154_arc['sample_depth'] = 140
jv154_bac['sample_depth'] = 140

# combine all data frames into one data set
combo = pd.concat([jv119_arc, jv119_bac, jv121_arc, jv121_bac, jv132_arc, jv132_bac, jv154_arc, jv154_bac])

# split the classification into tax level columns and rename
combo = split_classification(combo)

# write csv
#combo.to_csv(op.join(path,'tables/merged_mag_gtdb_table.csv'), index=False)
