==> old_notebooks/05_create_proximeta_assoc_gtdb_table.ipynb <==
# 05_create_proximeta_assoc_gtdb_table

Run this notebook to merge MAG GTDB data with proximeta associations.

# Load packages
import pandas as pd
import os
import os.path as op
import sys
import csv
import numpy as np
import matplotlib
import glob
import seaborn as sns
from collections import Counter
import matplotlib.pyplot as plt

# file paths
path = '/Users/melissaherring/Google Drive/My Drive/MH_project/'
#path = '/Users/juliabrown/Google Drive/My Drive/projects/OMZvir_round2/MH_project/'

# from 01_create_proximeta_associations_table:
proximeta_assoc = pd.read_csv(op.join(path, 'tables/proximeta_associations_table.csv'))

# from 03_create_merged_mag_gtdb_table:
mag_gtdb = pd.read_csv(op.join(path, 'tables/merged_mag_gtdb_table.csv'))

# rename mag_gtdb user_genome column to match associations data frame
mag_gtdb.rename(columns={'user_genome':'host_name'}, inplace=True) 

# Merge two data frames together
combo = proximeta_assoc.merge(mag_gtdb,how='left',on=["host_name","sample_name","sample_depth"])

# make all NA values match
combo = combo.fillna("NA")

# write csv
#combo.to_csv(op.join(path, 'tables/proximeta_assoc_gtdb_table.csv'), index=False)
