==> old_notebooks/08_create_vmag_assoc_summary_table.ipynb <==
# 08_create_vmag_assoc_summary_table

# Load packages
import pandas as pd
import os
import os.path as op
import sys
import csv
import numpy as np
import matplotlib
import glob
import seaborn as sns
from collections import Counter
import matplotlib.pyplot as plt

# file paths
path = '/Users/melissaherring/Google Drive/My Drive/MH_project/'
#path = '/Users/juliabrown/Google Drive/My Drive/projects/OMZvir_round2/MH_project/'

# from 01_create_proximeta_associations_table:
assoc = pd.read_csv(op.join(path, 'tables/proximeta_associations_table.csv'))

# filter out vMAGs only
assoc_vMAG = assoc[assoc['virus_type'] == 'vMAG']

# from proximeta viral files
vMAG_119 = pd.read_csv(op.join(path,'data/proximeta_viral_files/jv-119_proximeta/viral_MAGs/viral_mags_summary.tsv'), sep = '\t')
vMAG_121 = pd.read_csv(op.join(path,'data/proximeta_viral_files/jv-121_proximeta/viral_MAGs/viral_mags_summary.tsv'), sep = '\t')
vMAG_132 = pd.read_csv(op.join(path,'data/proximeta_viral_files/jv-132_proximeta/viral_MAGs/viral_mags_summary.tsv'), sep = '\t')
vMAG_154 = pd.read_csv(op.join(path,'data/proximeta_viral_files/jv-154_proximeta/viral_MAGs/viral_mags_summary.tsv'), sep = '\t')

# combine all vMAG data frames together
vMAGs_all = pd.concat([vMAG_119,vMAG_121,vMAG_132,vMAG_154])

# split up contig_id column into different columns
vMAGs_all['virus_name'] = vMAGs_all['contig_id'].str.split("|").str[0]
vMAGs_all['N'] = vMAGs_all['contig_id'].str.split("|").str[1]
vMAGs_all['N'] = vMAGs_all['N'].str.replace('N=', '')
vMAGs_all['N'] = pd.to_numeric(vMAGs_all['N'])
vMAGs_all['L'] = vMAGs_all['contig_id'].str.split("|").str[2]
vMAGs_all['L'] = vMAGs_all['L'].str.replace('L=', '')
vMAGs_all['L'] = pd.to_numeric(vMAGs_all['L'])

# merge vMAGs with associations 
combo = assoc_vMAG.merge(vMAGs_all,how='inner',on=['virus_name'])

combo.to_csv(op.join(path,'tables/vMAG_assoc_summary_table.csv'), index=False)
